[Refactoring️ ⚙<fe0f>] 경로 단축키로 변경
