+++ src/pages/Profile/Profile.tsx
@@ -1,5 +1,5 @@
-import UserPosts from "./Components/UserPosts/UserPosts"
-import UserProfile from "./components/UserProfile/UserProfile"
+import UserPosts from "@/pages/Profile/components/UserPosts/UserPosts"
+import UserProfile from "@/pages/Profile//components/UserProfile/UserProfile"
 
 const Profile = () => {
   return (
