+++ src/pages/Profile/components/UserProfile/UserNickname.tsx
@@ -1,7 +1,7 @@
-import * as S from "../../../../styles/Profile.styles"
+import * as GS from "@/styles/Profile.styles"
 
 const UserNickname = () => {
-  return <S.UserInfoContainer mb={1}>UserNickname</S.UserInfoContainer>
+  return <GS.UserInfoContainer mb={1}>UserNickname</GS.UserInfoContainer>
 }
 
 export default UserNickname
