+++ src/pages/Profile/components/UserProfile/UserProfileImage.tsx
@@ -1,4 +1,4 @@
-import * as S from "../../Styles/UserProfile.styles"
+import * as S from "@/pages/Profile/styles/UserProfile.styles"
 
 const UserProfileImage = () => {
   return <S.UserProfileImageContainer></S.UserProfileImageContainer>
